+++ analysis/src/main/java/cz/vutbr/fit/group/verifit/oslc/analysis/VeriFitAnalysisManager.java
@@ -117,9 +117,24 @@ public class VeriFitAnalysisManager {
     
     // Start of user code class_methods
 	
-	private static int getCurrentHighestAutomationRequestId() throws Exception
+	private static long getCurrentHighestAutomationRequestId() throws Exception
 	{
-    	int currMaxReqId = 0;
+		// first try to get the next ID from the bookmark resource
+		try {
+			AutomationRequest bookmarkReq = getAutomationRequest(null, "bookmarkID");
+			if (bookmarkReq != null)
+			{
+				String bookmarkIdentifier = bookmarkReq.getIdentifier();
+				if (bookmarkIdentifier != null)
+					return Long.parseLong(bookmarkIdentifier);
+			}
+		} catch (Exception e) {
+			
+		}
+		
+		// if the first method failed, fallback to the old slow method
+		log.warn("Initializing ID generators: bookmarkID resource not found. Falling back querying the whole database.");
+    	long currMaxReqId = 0;
     	try {
     		// loop over all the Automation Requests in the triplestore (TODO potential performance issue for initialization with a large persistent database)
     		for (int page = 0;; page++)
@@ -133,7 +148,7 @@ public class VeriFitAnalysisManager {
 				{	
 					for (AutomationRequest autoReq : listAutoRequests)
 					{
-						int reqId = Integer.parseInt(Utils.getResourceIdFromUri(autoReq.getAbout()));
+						long reqId = Long.parseLong(Utils.getResourceIdFromUri(autoReq.getAbout()));
 						if (reqId > currMaxReqId)
 						{
 							currMaxReqId = reqId;
@@ -145,6 +160,24 @@ public class VeriFitAnalysisManager {
 			throw new Exception(e.getMessage());
 		}
 		return currMaxReqId;
+	}
+	
+	private static void updateAutomationRequestIDbookmarkResource(String id)
+	{
+		AutomationRequest bookmarkReq = VeriFitAnalysisResourcesFactory.createAutomationRequest("bookmarkID");
+		bookmarkReq.setIdentifier(id);
+		bookmarkReq.setDescription("This resource is used by the adapter to rememeber what is the current maximum AutomationRequest ID. "
+				+ "Determines where new ID's start after adapter restart with a persistent triplestore.");
+		
+        Store store = storePool.getStore();
+        try {
+            store.updateResources(storePool.getDefaultNamedGraphUri(), bookmarkReq);
+        } catch (StoreAccessException e) {
+            log.error("Failed to create resource: '" + bookmarkReq.getAbout() + "'", e);            
+            throw new WebApplicationException("Failed to create resource: '" + bookmarkReq.getAbout() + "'", e, Status.INTERNAL_SERVER_ERROR);
+        } finally {
+            storePool.releaseStore(store);
+        }
 	}
 	
 	/**
@@ -604,7 +637,7 @@ public class VeriFitAnalysisManager {
 		}
 
 		// check what the last AutomationRequest ID is (requests have a numerical ID) and initialize the ID generator to one higher
-    	int initReqId = 0;
+    	long initReqId = 0;
 		try {
 			initReqId = getCurrentHighestAutomationRequestId() + 1;
 		} catch (Exception e) {
@@ -732,6 +765,17 @@ public class VeriFitAnalysisManager {
         // End of user code
         
         // Start of user code queryAutomationRequests
+        
+        for (AutomationRequest req : resources)
+        {
+            // dont show the bookmark resource in query responses
+            if (Utils.getResourceIdFromUri(req.getAbout()).equals("bookmarkID"))
+            {
+            	resources.remove(req);
+            	break;
+            }
+        }
+        
         // End of user code
         return resources;
     }
@@ -845,9 +889,13 @@ public class VeriFitAnalysisManager {
 				throw new OslcResourceException("title property missing");
 			
 			
+			// generate a new ID and update the ID bookmark resource 
+			String newID = AutoRequestIdGen.getId();
+			updateAutomationRequestIDbookmarkResource(newID);
+			
+			
 			// create a basic Automation Request (with core properties like creation time, identifier, ...)
 			// and set relevant properties from the Automation Request received from the client
-			String newID = AutoRequestIdGen.getId();
 			newResource = VeriFitAnalysisResourcesFactory.createAutomationRequest(newID);
 			newResource.setInputParameter(aResource.getInputParameter());
 			newResource.setTitle(aResource.getTitle());
@@ -857,7 +905,6 @@ public class VeriFitAnalysisManager {
 			newResource.setContributor(aResource.getContributor());
 			newResource.setExtendedProperties(aResource.getExtendedProperties());
 			//newResource.replaceDesiredState(aResource.getDesiredState());	// TODO use this to implement deferred execution later
-			
 			
 			// get the executed autoPlan and load its configuration
 			String execAutoPlanId = Utils.getResourceIdFromUri(newResource.getExecutesAutomationPlan().getValue());
@@ -1009,6 +1056,14 @@ public class VeriFitAnalysisManager {
         Boolean deleted = false;
         // Start of user code deleteAutomationRequest_storeInit
 
+        // dont allow clients to delete the bookmark resource
+        if (id.equals("bookmarkID"))
+        {
+            log.error("Automation Request DELETE: bookmarkID resource is not allowed to be updated");
+            throw new WebApplicationException("Automation Request DELETE: bookmarkID resource is not allowed to be deleted", Status.FORBIDDEN);
+        }
+        
+
         AutomationRequest requestToDelete = null;
         try {
         	requestToDelete = getAutomationRequest(null, id);
@@ -1042,7 +1097,7 @@ public class VeriFitAnalysisManager {
         
         // Start of user code deleteAutomationRequest
         if (httpServletRequest != null) {
-	        String cascade = httpServletRequest.getHeader("cascade");
+	        String cascade = httpServletRequest.getParameter("cascade");
 	        if (cascade != null && cascade.equalsIgnoreCase("true"))
 	        {
 	        	try {
@@ -1060,6 +1115,14 @@ public class VeriFitAnalysisManager {
         AutomationRequest updatedResource = null;
         // Start of user code updateAutomationRequest_storeInit
 
+        // dont allow clients to update the bookmark resource
+        if (id.equals("bookmarkID"))
+        {
+            log.error("Automation Request UPDATE: bookmarkID resource is not allowed to be updated");
+            throw new WebApplicationException("Automation Request UPDATE: bookmarkID resource is not allowed to be updated", Status.FORBIDDEN);
+        }
+        
+        // check that the request has all the required properties
         if (aResource == null)
         {
             log.error("Automation Request UPDATE: received an empty request");
@@ -1218,7 +1281,7 @@ public class VeriFitAnalysisManager {
         
         // Start of user code deleteAutomationResult
         if (httpServletRequest != null) {
-	        String cascade = httpServletRequest.getHeader("cascade");
+	        String cascade = httpServletRequest.getParameter("cascade");
 	        if (cascade != null && cascade.equalsIgnoreCase("true"))
 	        {
 	        	try {