+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### v2.2.0
 - implemented resource delete
   - delete capabilities for A.Requests, A.Results, Contributions (analysis only), SUT (compilation only)
-  - Requests, Results, and SUTs recognize a "cascade" header which when set to "true" makes it so that 
+  - Requests, Results, and SUTs recognize a "cascade" parameter which when set to "true" makes it so that 
     deleting any one of the three will result in all three getting deleted
   - deleting a resource that is connected to a directory or a file will also delete the file/directory (SUT, contribution, A.Result)
   - deleting an Automation Request that is not yet finished results in the request getting canceled first (using desiredState update),
@@ -13,7 +13,9 @@
 - implemented execution cancel
   - cancel Automation Request execution by updating the A.Request's desiredState property to "canceled"
   - canceling only works on requests that are still state "inProgress"
-
+- reworked resource ID generators
+  - now use long instead of int
+  - initialization is based on a special bookmarkID resource instead of walking the whole triplestore
 
 ### v2.1.1
 - fixed high cpu usage on Windows